backend-src/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
tweepy==4.14.0
google-analytics-data==0.16.2
hubspot-api-client==7.3.0
sendgrid==6.10.0
stripe==7.6.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
pandas==2.1.3
scikit-learn==1.3.2
requests==2.31.0
httpx==0.24.0
